Refactor: Move DdmAdminClient wrapper out of sled-agent/installinator into ddm-admin-client crate.
