+++ ddm-admin-client/Cargo.toml
@@ -8,6 +8,11 @@ progenitor-client.workspace = true
 reqwest = { workspace = true, features = ["json", "stream", "rustls-tls"] }
 serde.workspace = true
 slog.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+
+omicron-common.workspace = true
+sled-hardware.workspace = true
 
 [build-dependencies]
 anyhow.workspace = true
