+++ installinator/Cargo.toml
@@ -23,7 +23,6 @@ ipcc-key-value.workspace = true
 itertools.workspace = true
 once_cell.workspace = true
 omicron-common.workspace = true
-omicron-sled-agent.workspace = true
 progenitor-client.workspace = true
 reqwest.workspace = true
 serde.workspace = true
