+++ sled-agent/src/bootstrap/mod.rs
@@ -7,7 +7,6 @@
 pub mod agent;
 pub mod client;
 pub mod config;
-pub mod ddm_admin_client;
 mod hardware;
 mod http_entrypoints;
 mod maghemite;
